ai-notebook/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
sqlalchemy==1.4.53
cryptography==41.0.7
Pillow==10.0.1
python-magic==0.4.27
python-magic-bin==0.4.14